backendAPI/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
torch==2.1.2
torchvision==0.16.2
transformers==4.35.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
aiohttp==3.9.1
python-multipart==0.0.6
